subo cambios restantes y arreglo de cache

+++ src/main/java/com/meli/solr/api/service/SolrSysService.java
@@ -8,6 +8,6 @@ public interface SolrSysService {
 
 	void init();
 
-	Map<WeatherType, Integer> getReport();
+	Map<WeatherType, Long> getReport();
 
 }
